Add alternate engine implementation
This adds an engine for fiahod. This engine is trimmed down from the more
cinematic original engine because while writing it I ran into issues
that I chalked up to the engine being too complex. This might not
actually be the case. Anyway, it sounds a lot different from the
original implementation and is its own more analog flavor.

It has some issues, though:
- It clips quite a bit
- The parameters are pretty jumpy and chaotic
- The "dust" seems to be broken somehow
- The reverb isn't smudgy enough and feels too rhythmic and echo-y

I'm posting this code up because I like the way this sounds, but I
think the original implementation would be a better starting point.